feat: move logic to a hook to fill the ranking
